+++ postcss.config.js
@@ -5,6 +5,7 @@ const colors = {
   'error-text-color': '#ff4d00',
   'text-color-light': '#999999',
   'primary-bg-color': 'rgb(255 255 255 / 90%)',
+  'secondary-btn-bg-color': 'rgb(255 255 255 / 40%)',
   'disabled-bg-color': '#999999',
   'accent-color': '#ffc701',
   'pure-white': '#ffffff',
