+++ src/components/Button/button.css
@@ -42,7 +42,7 @@
 
   &_view_secondary {
     border: 1px solid $primary-color;
-    background: $primary-bg-color;
+    background: $secondary-btn-bg-color;
     color: $primary-color;
     display: flex;
     justify-content: center;
