+++ src/pages/ProfilePage/Header/Header.css
@@ -1,0 +1,2 @@
+.header {
+}
